makeDesign

This function will aid in making a design file for an imaging experiment (design.csv).

Usage

makeDesign(
  proj,
  n.plate,
  n.row = 8,
  n.col = 12,
  assay.type = NULL,
  food = NULL,
  od = NULL
)

Arguments

proj

The name of the project this design file will be used for. Remember that the project name always follows the 8-digit date in project directories. For example, the project directory 20200130_toxin01A has a project name of "toxin01A". The project name will be used to join the design file to the cellprofiler-nf output data by the Metadata_Experiment variable, so it is critical that they match exactly.

n.plate

The number of plates in the experiment.

n.row

The number of rows in the plates. Default is 8 for a 96-well plate.

n.col

The number of columns in the plates. Default is 12 for a 96-well plate.

assay.type

OPTIONAL: The type of assay run. For example, "48h".

food

OPTIONAL: A character description of the food used to perform the assay. For example, "15hHB101_20220727".

od

OPTIONAL: The final optical density of bacterial food used in the assay. For example, 10.

Value

A single data frame with the variables "Metadata_Plate" and "Metadata_Well" variables that can be used as a design file for the readXpress function. The data frame can be augmented further to include specific drugs and dilutents or other experimental design attributes if desired. The variables output include:

Details

The variables output are as follows:

Variables	Class

Metadata_Experiment	chr

Metadata_Plate	chr

Metadata_Well	chr

strain	chr

drug	chr

concentration_um	dbl

bleach	chr

diluent	chr

assay_type	chr

plate	dbl

row	chr

col	chr

food	chr

od	int

well_censor	chr

well_censor_reason	chr

notes	chr